m1_ai_app/backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
